# docs/swagger.yaml
basePath: /
definitions:
  github_com_Nezzy-joe_task-api_models.Task:
    properties:
      completed:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
  handlers.authCredentialsRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.authResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/handlers.authUserResponse'
    type: object
  handlers.authUserResponse:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: A secure Task Management API built with Go and Supabase Auth.
  title: Task API
  version: "2.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user through Supabase Auth and returns access and
        refresh tokens.
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.authCredentialsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.authResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Log in a user
      tags:
      - Authentication
  /auth/logout:
    post:
      description: Revokes the authenticated user's refresh tokens through Supabase
        Auth.
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Log out the authenticated user
      tags:
      - Authentication
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user through Supabase Auth.
      parameters:
      - description: Signup credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.authCredentialsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.authResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Authentication
  /protected/dashboard:
    get:
      description: Demonstrates reuse of the authentication middleware on a second
        protected route.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Access protected dashboard
      tags:
      - Protected
  /protected/profile:
    get:
      description: Returns safe metadata for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get authenticated profile
      tags:
      - Protected
  /public/info:
    get:
      description: Returns information that does not require authentication.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get public information
      tags:
      - Public
  /tasks:
    get:
      description: Return all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Nezzy-joe_task-api_models.Task'
            type: array
      summary: Get all tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new task
      parameters:
      - description: Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/github_com_Nezzy-joe_task-api_models.Task'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Nezzy-joe_task-api_models.Task'
        "400":
          description: Invalid JSON
          schema:
            type: string
      summary: Create a new task
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      description: Delete a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid task ID
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            type: string
      summary: Delete a task
      tags:
      - Tasks
    get:
      description: Return a single task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Nezzy-joe_task-api_models.Task'
        "400":
          description: Invalid task ID
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            type: string
      summary: Get task by ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update an existing task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/github_com_Nezzy-joe_task-api_models.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Nezzy-joe_task-api_models.Task'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            type: string
      summary: Update a task
      tags:
      - Tasks
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token as: Bearer <JWT>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"